pyproject.toml
[tool.poetry]
name = "Algotard"
version = "0.1.0"
description = "A cryptocurrency trading bot using RSI strategy."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
packages = [{ include = "algotard", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
ccxt = "^3.1.6"
pandas = "^2.0.3"
ta = "^0.10.2"
jupyter = "^1.1.1"
matplotlib = "^3.9.2"
mailjet-rest = "^1.3.4"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
isort = "^5.13.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"